Updating graph interfaces
