+++ src/spurt/graph/__init__.py
@@ -1,5 +1,6 @@
-from ._delaunay import DelaunayGraph, Reg2DGraph, order_points
+from ._delaunay import DelaunayGraph, Reg2DGraph
 from ._interface import GraphInterface, PlanarGraphInterface
+from .utils import order_points
 
 __all__ = [
     "order_points",
